timeslice parameter now is a double, not an integer

+++ src/main/scala/utils/slicer/time/TimeSlicer.scala
@@ -6,7 +6,7 @@ import org.joda.time.DateTime
 trait TimeSlicer extends Serializable {
   val conf = ConfigFactory.load()
 
-  def getSliceForTimestamp(timestampInCsv: String, sliceSize: Int): Int
+  def getSliceForTimestamp(timestampInCsv: String, sliceSize: Double): Int
 
   def getTimestampForSlice(slice: Int, sliceSize: Int): DateTime
